+++ lib/dynomite/reserved_words.rb
@@ -1,0 +1,18 @@
+module Dynomite
+  RESERVED_WORDS = %w[
+    as_json
+    attrs
+    attributes
+    delete
+    columns
+    find
+    getter
+    new_record
+    param_name
+    partition_key
+    replace
+    setter
+    scan
+    table_name
+  ].freeze
+end