+++ CHANGELOG.md
@@ -3,6 +3,13 @@
 All notable changes to this project will be documented in this file.
 This project *tries* to adhere to [Semantic Versioning](http://semver.org/), even before v1.0.
 
+## [UNRELEASED]
+- Add a way to quickly define getters and setters using `column` method
+- Can be used with `ActiveModel::Validations`
+- Add ActiveModel::Validations (group=test,development) dependency
+- Add ActiveModel::Validations Item integration spec
+- Add Dynomite::Item.replace and .replace! spec with validations
+
 ## [1.1.1]
 - #6 from patchkit-net/feature/table-count: add Item.count